+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+### Breaking changes
+- Change how options should be provided to the `fcom` command
+- Parse command argument/options with `slop` gem
+
 ## 0.1.0 - 2019-12-28
 ### Added
 - Initial release